Max Datom - level 10: seed users data and complete the level.
There's new schema-users.edn and data-users.edn.
The schema was extracted supplied in the level definition:
https://max-datom.com/#/DB612D03-9AF7-49B7-98B5-4C77ADE029D2
The data were extracted with the queries (see data-users.edn and level_10.clj).

The task was confusing - they said:
> pdate the query to return the :user/first+last-name
> and **count of :post/author** for the user associated with :user/id #uuid "1B341635-BE22-4ACC-AE5B-D81D8B1B7678"

But instead, it seems it must be count of `:post/id`,
otherwise it returns 1 (it's obviously only 1 distict author)
instead of 3 (number of posts wrote by that author).